스프링 시큐리티 쉽게 이해

+++ Chapter06_Thymeleaf/src/main/java/edu/pnu/controller/BoardController.java
@@ -1,4 +1,4 @@
-package edu.pnu.controller;
+package edu.pnu.controller; //실제 이런 코드는 없고 필터를 사용해서 인증
 
 import java.util.List;
 
@@ -6,23 +6,37 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
 import edu.pnu.domain.Board;
+import edu.pnu.domain.Member;
 import edu.pnu.service.BoardService;
 
+@SessionAttributes("member") //세션에 상태 정보 저장할 때 사용 > Model에 member라는 이름으로 저장된 데이터를 자동으로 세션에 등록
 @Controller
 public class BoardController {
 	
 	@Autowired
 	private BoardService boardService;
 	
+	@ModelAttribute("member") //Model 객체에다가 return하여 생성된 객체를 member라는 이름으로 저장해라 > save! 로그인하든 하지 않든 비어있는 객체 생성 + 초기화
+	public Member setMember() {
+		return new Member();
+	}
+	
 	@GetMapping("/getBoardList")
-	public String getBoardList(Model model, Board board) {
+	public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) { //세션에 저장된 member 객체를 찾아서 넣어줌
+		
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		List<Board> boardList = boardService.getBoardList(board);
 		
-		model.addAttribute("boardList",boardList);
+		model.addAttribute("boardList", boardList);
 		return "getBoardList";
 	}
 //	@GetMapping("/getBoardList")
@@ -42,41 +56,63 @@ public class BoardController {
 //	}
 	
 	@GetMapping("/insertBoard")
-	public String insertBoardView() {
+	public String insertBoardView(@ModelAttribute("member") Member member) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		return "insertBoard";
 	}
 	
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 	}
 	
-//	@GetMapping("/getBoard")
-//	public String getBoard(Board board, Model model) {
-//		model.addAttribute("board", boardService.getBoard(board));
-//		return "getBoard"; //view name return
-//	}
-	
 	@GetMapping("/getBoard")
-	public ModelAndView getBoard(Board board) {
-		Board b = boardService.getBoard(board);
+	public String getBoard(@ModelAttribute("member") Member member, Board board, Model model) {
 		
-		ModelAndView mv = new ModelAndView();
-		mv.addObject("board", b);
-		mv.setViewName("getBoard");
-		return mv;
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
+		model.addAttribute("board", boardService.getBoard(board));
+		return "getBoard"; //view name return
 	}
 	
+//	@GetMapping("/getBoard")
+//	public ModelAndView getBoard(Member member, Board board) {
+//		Board b = boardService.getBoard(board);
+//		
+//		ModelAndView mv = new ModelAndView();
+//		mv.addObject("board", b);
+//		mv.setViewName("getBoard");
+//		return mv;
+//	}
+	
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.updateBoard(board);
 		return "redirect:getBoardList";
 	}
 	
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
+	
 }
